test(core): migrate TerminalLaunchEnvironmentTests (Wave D)
1 test, 5 failure sites preserved 1:1.
